# Improve SelectVariant performance