--- a/src/Unleash/Variants/VariantUtils.cs
+++ b/src/Unleash/Variants/VariantUtils.cs
@@ -12,28 +12,30 @@ namespace Unleash.Variants
 
         public static Variant SelectVariant(string groupId, UnleashContext context, List<VariantDefinition> variantDefinitions)
         {
-            var stickiness = variantDefinitions[0].Stickiness ?? "default";
+            var stickiness = variantDefinitions.FirstOrDefault()?.Stickiness ?? "default";
             var target = StrategyUtils.GetNormalizedNumber(GetIdentifier(context, stickiness), groupId, VARIANT_NORMALIZATION_SEED);
 
             var counter = 0;
+            Variant result = null;
             foreach (var variantDefinition in variantDefinitions)
             {
                 if (variantDefinition.Weight != 0)
                 {
                     if (variantDefinition.Overrides.Count > 0 && variantDefinition.Overrides.Any(OverrideMatchesContext(context)))
                     {
-                        return variantDefinition.ToVariant();
+                        result = variantDefinition.ToVariant();
+                        break;
                     }
 
                     counter += variantDefinition.Weight;
-                    if (counter >= target)
+                    if (counter >= target && result == null)
                     {
-                        return variantDefinition.ToVariant();
+                        result = variantDefinition.ToVariant();
                     }
                 }
             }
 
-            return null;
+            return result;
         }
 
         public static Variant SelectVariant(FeatureToggle feature, UnleashContext context, Variant defaultVariant = null)